backend/requirements.txt
# Flask 및 웹 프레임워크
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
werkzeug>=2.3.0

# OpenAI 및 LLM
openai>=1.0.0
requests>=2.31.0

# 데이터 처리
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# RAG 및 벡터 데이터베이스
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0
sentence-transformers>=2.2.0

# 머신러닝
torch>=2.0.0
transformers>=4.30.0

# 한국어 처리
jieba>=0.42.1 